feat(frontend): Add tailwind.css to the project
Add tailwind.css to the project. At the moment, everything must manually be
installed using NPM, and you must run the scripts manually to build app.min.css.
For example, run npm run watch to run tailwind, checking for changes. In the
future, this will be setup to be executed automatically on build / debug. See
https://stackoverflow.com/questions/57669027/how-do-you-add-tailwind-css-into-a-
blazor-app for information on how to possibly do this.