feat(frontend): Implement Darkmode toggle
Implement a darkmode toggle based on the instructions given here:
https://jonhilton.net/blazor-tailwind-dark-mode-local-storage/. This uses an
injected IProfileService in both the client and server. The server side is a
dummy implementation temporarily, just to satisfy prerendering dependencies. The
client implementation uses the BlazoredLocalStorage extension to access local
storage. In the future, a better system such as database stored preferences will
be implemented.